+++ src/frontend/src/features/recording/components/NoAccessView.tsx
@@ -79,11 +79,20 @@ export const NoAccessView = ({
           })}
         >
           {t(`${i18nKey}.body`)}
-          <br />
           {helpArticle && (
-            <A href={helpArticle} target="_blank">
-              {t(`${i18nKey}.linkMore`)}
-            </A>
+            <>
+              {' '}
+              <A
+                href={helpArticle}
+                target="_blank"
+                rel="noopener noreferrer"
+                externalIcon
+                color="note"
+                aria-label={t(`${i18nKey}.linkAriaLabel`)}
+              >
+                {t(`${i18nKey}.linkMore`)}
+              </A>
+            </>
           )}
         </Text>
       </VStack>